pyproject.toml
[project]
requires-python = "==3.13.*"
name = "cancer"
description = "Cures cancer"
version = "1.0.0"
authors = [{ name = "BlindfoldedSurgery", email = "team@blindfolded.surgery" }]
classifiers = [
    "Private :: Do Not Upload",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "bs-config [dotenv] ==1.1.1",
    "click >=8.1.3, <9.0.0",
    "google-cloud-pubsub >=2.0.0, <3.0.0",
    "httpx ==0.28.*",
    "Pillow >=11.0.0, <12.0.0",
    "python-telegram-bot ==22.0",
    "sentry-sdk >=2.0.0, <3.0.0",
    "yt-dlp [default] ==2025.4.30",
]

[dependency-groups]
dev = [
    "mypy ==1.15.*",
    "pytest >=8.0.0, <9.0.0",
    "ruff ==0.11.9",
    "types-requests >=2.28.11, <3.0.0",
    "types-Pillow >=10.0.0, <11.0.0",
]

[[tool.uv.index]]
name = "internal"
url = "https://pypi.bjoernpetersen.net"
explicit = true

[tool.uv.sources]
bs-config = { index = "internal" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yt_dlp.*"
ignore_missing_imports = true

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "FURB",
    "G",
    "I001",
    "LOG",
    "PTH",
    "Q",
    "UP",
]
ignore = [
    # redundant-open-modes
    "UP015",
]
